refactor(userspace): extract CoS config snapshots to protocol_cos.go
Second code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the class-of-service configuration snapshot cluster into a
cohesive sibling protocol_cos.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change):
ClassOfServiceSnapshot, CoSForwardingClassSnapshot,
CoSDSCPClassifierSnapshot, CoSDSCPClassifierEntrySnapshot,
CoSIEEE8021ClassifierSnapshot, CoSIEEE8021ClassifierEntrySnapshot,
CoSDSCPRewriteRuleSnapshot, CoSDSCPRewriteRuleEntrySnapshot,
CoSSchedulerSnapshot, CoSSchedulerMapSnapshot,
CoSSchedulerMapEntrySnapshot.

These are the Go->Rust CoS config wire types; every json tag is
preserved byte-identical. The CoS runtime *status* types stay elsewhere
and move in a later increment. Build passes; gofmt clean.
